Implement delete method to handle exercises and workouts removal; implement post, patch, get and delete methods in ActivityView

+++ workouts/views.py
@@ -3,10 +3,11 @@ from rest_framework.response import Response
 from rest_framework import status
 from django.contrib.auth import get_user_model
 from django.db.models import Q
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
+from datetime import datetime
 
-from .serializers import ExerciseSerializer, WorkoutSerializer, BlockSerializer, BlockExerciseSerializer, BlockExerciseDataSerializer
-from .models import Exercise, Workout
+from .serializers import ExerciseSerializer, WorkoutSerializer, BlockSerializer, BlockExerciseSerializer, BlockExerciseDataSerializer, ActivitySerializer
+from .models import Exercise, Workout, Activity
 
 User = get_user_model()
 
@@ -28,6 +29,19 @@ class ExercisesList(APIView):
 
         return Response(serializer.data, status=status.HTTP_200_OK)
     
+    def delete(self, request, exercise_id=None):
+        try:
+            target_exercise = Exercise.objects.get(id=exercise_id)
+        except Exercise.DoesNotExist:
+            raise NotFound({"detail": "Exercise not found."})
+        
+        if target_exercise.user != request.user:
+            raise PermissionDenied({"detail": "You do not have permission to delete this exercise."})
+    
+        target_exercise.delete()
+        return Response({}, status=status.HTTP_204_NO_CONTENT)
+
+    
 class WorkoutsList(APIView):
     def post(self, request):
         new_workout = request.data
@@ -44,7 +58,7 @@ class WorkoutsList(APIView):
         # Create blocks
         for i, block in enumerate(new_workout["blocks"]):
             new_block = {
-                "workout": workout_serializer.data["id"],
+                "workout": workout_serializer.instance.id,
                 "position": i
             }
 
@@ -58,7 +72,7 @@ class WorkoutsList(APIView):
             # Create block exercises
             for i, exercise in enumerate(block["exercises"]):
                 block_exercise = {
-                    "block": block_serializer.data["id"],
+                    "block": block_serializer.instance.id,
                     "exercise": exercise["exercise"],
                     "position": i
                 }
@@ -68,11 +82,11 @@ class WorkoutsList(APIView):
                 if block_exercise_serializer.is_valid():
                     block_exercise_serializer.save()
                 else:
-                    return Response(block_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                    return Response(block_exercise_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
                 for field in exercise["fields"]:
                     new_entry = {
-                        "block_exercise": block_exercise_serializer.data["id"],
+                        "block_exercise": block_exercise_serializer.instance.id,
                         "field": field,
                         "value": exercise["data"][field]
                     }
@@ -115,9 +129,12 @@ class WorkoutsList(APIView):
     def delete(self, request, workout_id=None):
         try:
             target_workout = Workout.objects.get(id=workout_id)
-        except Exercise.DoesNotExist:
+        except Workout.DoesNotExist:
             raise NotFound({"detail": "Workout not found."})
         
+        if target_workout.user != request.user:
+            raise PermissionDenied({"detail": "You do not have permission to delete this workout."})
+    
         target_workout.delete()
         return Response({}, status=status.HTTP_204_NO_CONTENT)
     
@@ -182,3 +199,54 @@ class WorkoutsList(APIView):
                         return Response(block_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
         return Response(workout_serializer.data, status=status.HTTP_202_ACCEPTED)
+    
+
+class ActivityView(APIView):
+    def post(self, request):
+        activity_entry = request.data
+        activity_entry["user"] = request.user.id
+
+        serializer = ActivitySerializer(data=activity_entry)
+
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+    def patch(self, request, activity_id=None):
+        data_update = request.data
+      
+        try:
+            target_instance = Activity.objects.get(pk=activity_id)
+        except Activity.DoesNotExist:
+            raise NotFound({"detail": "Activity entry not found."})
+
+        serializer = ActivitySerializer(target_instance, data=data_update, partial=True)
+
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        
+    def delete(self, request, activity_id=None):
+        try:
+            target = Activity.objects.get(pk=activity_id)
+        except Activity.DoesNotExist:
+            raise NotFound({"detail": "Activity entry not found."})
+        
+        if target.user != request.user:
+            raise PermissionDenied({"detail": "You do not have permission to delete this entry."})
+    
+        target.delete()
+        return Response({}, status=status.HTTP_204_NO_CONTENT)
+    
+    def get(self, request):
+        try:
+            activity = request.user.fitness_activity.order_by("-date_scheduled")
+            serializer = ActivitySerializer(activity, many=True)
+
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        except Activity.DoesNotExist:
+            return Response({"details": "No activity found for this user."})
